refactor(world): delete bridge, revert BaseWidget/useUpstreamValue
Phase 2b of temp/plans/world-consolidation.md (completes Strategy A
bridge elimination begun in Phase 2a).

Now that useWidgetValueStore is a World facade (Phase 2a), the bridge
is redundant.

Deletions:
- src/world/widgetWorldBridge.ts + widgetWorldBridge.test.ts deleted.
  Their register/getNodeWidgets coverage rolled into widgetValueStore
  facade tests in Phase 2a; widgetParent tests already moved to
  src/stores/widgetComponents.test.ts in Phase 1.

Reverts (back to pre-slice-1 baseline):
- BaseWidget.setNodeId no longer double-writes via
  registerWidgetInWorld; the store IS the World writer now. Drops
  three @/world/* imports.
- useUpstreamValue reads via useWidgetValueStore().getNodeWidgets().
  Drops three @/world/* imports.

Test updates:
- useUpstreamValue.test.ts setup uses
  useWidgetValueStore().registerWidget instead of
  registerWidgetInWorld(getWorld(), ...). Hoists
  setActivePinia(createTestingPinia) + resetWorldInstance into
  beforeEach.
- widgetComponents.test.ts setup inlines a 4-line widget registration
  to replace the deleted bridge import.
- entityIds.ts: GraphId type un-exported (no external consumer; YAGNI;
  re-export when slice 2 needs it).

End state:
- src/world/ is pure substrate (5 source files: brand, entityIds,
  componentKey, world, worldInstance). No bridge, no components dir.
- BaseWidget.ts byte-identical with pre-slice-1 form at the
  setNodeId seam.
- useWidgetValueStore is the sole owner of widget value state;
  Vue tracking flows naturally through reactive(Map) inside the World.

Verification:
- pnpm typecheck, format:check, knip clean.
- 52 tests pass across src/world, src/stores/widgetValueStore,
  src/stores/widgetComponents, src/composables/useUpstreamValue,
  src/lib/litegraph/src/widgets/BaseWidget.
- rg "@/world" src/lib/litegraph/src/widgets/BaseWidget.ts returns 0.
- rg "@/world" src/composables/useUpstreamValue.ts returns 0.
- rg "registerWidgetInWorld|getNodeWidgetsThroughWorld|widgetWorldBridge" src/ returns 0.

Combined Phase 2 non-test diff -53 LOC (under 150 LOC budget).
BaseWidget._state shared reactive identity contract preserved end-to-end.

Amp-Thread-ID: https://ampcode.com/threads/T-019dd146-a3ad-734d-9825-0ab356454dd5
