+++ src/stores/widgetComponents.test.ts
@@ -2,10 +2,14 @@ import { describe, expect, it } from 'vitest'
 
 import type { WidgetState } from '@/stores/widgetValueStore'
 import { asGraphId, nodeEntityId, widgetEntityId } from '@/world/entityIds'
-import { registerWidgetInWorld } from '@/world/widgetWorldBridge'
 import { createWorld } from '@/world/world'
 
-import { widgetParent } from './widgetComponents'
+import type { WidgetValue } from './widgetComponents'
+import {
+  WidgetContainerComponent,
+  WidgetValueComponent,
+  widgetParent
+} from './widgetComponents'
 
 const graphId = asGraphId('00000000-0000-0000-0000-000000000001')
 
@@ -16,9 +20,14 @@ function makeState(nodeId: string, name: string, value: unknown): WidgetState {
 describe('widgetParent', () => {
   it('returns the owning node entity for a widget', () => {
     const world = createWorld()
-    registerWidgetInWorld(world, graphId, makeState('node-1', 'seed', 1))
-    const widgetId = widgetEntityId(graphId, 'node-1', 'seed')
-    expect(widgetParent(world, widgetId)).toBe(nodeEntityId(graphId, 'node-1'))
+    const state = makeState('node-1', 'seed', 1)
+    const widgetId = widgetEntityId(graphId, state.nodeId, state.name)
+    const ownerId = nodeEntityId(graphId, state.nodeId)
+    world.setComponent(widgetId, WidgetValueComponent, state as WidgetValue)
+    world.setComponent(ownerId, WidgetContainerComponent, {
+      widgetIds: [widgetId]
+    })
+    expect(widgetParent(world, widgetId)).toBe(ownerId)
   })
 
   it('returns undefined when no container references the widget', () => {
